os-* ciao-*: Use `service`_fqdn for reference of a node.
There are some places (like certificates) where the
controller hostname should be used.

This was achieved with groups['ciao_controller'][0] which worked
when the host file was defined as:

[ciao_controller]
controller.example.com

But if the IP is defined in the host file instead of the fqdn,
the ip was getting used instead of its fqdn.

- Replace {{ groups['ciao_controller'][0] }}
  with {{ ciao_controller_fqdn }}

- Replace {{ groups['openstack_identity'][0] }}
  with {{ keystone_fqdn }}

- Replace {{ groups['openstack_object_storage_controller'][0] }}
  with {{ swift_fqdn }}

Signed-off-by: Alberto Murillo <alberto.murillo.silva@intel.com>