examples: ciao: Include ciao_dev var in examples
If ciao_dev is set to false (default) it will try to install
ciao from the operating system packages which is only supported
in Clear Linux.

In order to install ciao in Ubuntu, ciao_dev must be set to
true to build ciao from github master branch.

This commit adds the variable to the examples to have it handy
for users who wants to deploy ciao in Ubuntu.

Signed-off-by: Alberto Murillo <alberto.murillo.silva@intel.com>