ciao-controller: Create the cnci image once.
This commit creates /var/lib/ciao/images/.cnci_ready file after
it has finished with the configuration of a cnci image.

If the file exists then the cnci image setup will be skipped.

This change prevents the playbook from creating and transfering the cnci
image each time is run and only does it when its necessary.

Signed-off-by: Alberto Murillo <alberto.murillo.silva@intel.com>