ciao: create admin-openrc and demo-openrc
The playbooks were creating a ciaorc to authenticate the cli
with the ciao user.

The ciao user is reserved for ciao services authenticating into keystone
and should not be used in the cli to create instances.

This changes creates admin-openrc with credentials for the admin user
and demo-openrc with credentials for the demo user to be used by ciao-cli.

Fixes #46

Signed-off-by: Alberto Murillo <alberto.murillo.silva@intel.com>