ciao: Change default value for gopath
Using $GOPATH as the value for gopath is not
so easily achievable because the GOPATH value
for the deployment machine does not get copied
automatically to other nodes.

Also setting a fact on one node does not set it
on other nodes.

The easiest way to share this value across the
different ciao roles is to set it as a variable
that can be overriden in groups_vars/all file or
via the cmdline with --extra-vars "gopath=$GOPATH"

Fixes: #57

Signed-off-by: Alberto Murillo <alberto.murillo.silva@intel.com>