ciao-compute/ciao-network: Ensure qemu processes stay alive
When restarting ciao-launcher on the CN or NN it's important to keep the
qemu processes that they've started alive.

The default, KillMode=control-group, will result in all processes in
the control group being killed when the service is stopped. This will
include all the qemu processes created by ciao-launcher as they are
placed into the same cgroup by systemd.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>